added method MathUtils.digits
Notes on performance: I compared this to two other version: a one-liner (`[...\`${num}\`].map(n => Number.parseInt(n))`) and a version that was similar to the version in this commit except that it used BigInts for the integer division instead of rounding down. I tested these on calculating the digits of all the numbers from 0 to 10^7.

Results: the algorithm in this commit took 4 seconds on average, the one-liner took 7 seconds on average, and the version with BigInts took 14 seconds on average.